+++ test/integ/descriptor/__init__.py
@@ -5,7 +5,7 @@ Integration tests for stem.descriptor.* contents.
 __all__ = [
   'extrainfo_descriptor',
   'microdescriptor',
+  'networkstatus',
+  'remote'
   'server_descriptor',
-  'get_resource',
-  'open_desc',
 ]